Add ForcedOperation.FatalError() and use to replace oft replicated error checking sequence where possible.
